feat: added support for non event sourced aggregates.
changes:
- created new class which not follow versions and other stuff specific for event sourcing pattern
- moved and renamed some classes from domain into domain events project
- added DeletebleObject.cs to allow to only soft delete objects from the system
- created interfaces for aggregate root classes and entities
