+++ Examples/EssentialFrame.ExampleApp/EssentialFrame.ExampleApp.csproj
@@ -12,9 +12,11 @@
         <ProjectReference Include="..\..\Src\Commands\Validations\EssentialFrame.Cqrs.Commands.Validations\EssentialFrame.Cqrs.Commands.Validations.csproj"/>
         <ProjectReference Include="..\..\Src\Core\Tests\EssentialFrame.Tests.UnitTests\EssentialFrame.Tests.UnitTests.csproj"/>
         <ProjectReference Include="..\..\Src\DomainEvents\EssentialFrame.Domain.Events\EssentialFrame.Domain.Events.csproj"/>
-        <ProjectReference Include="..\..\Src\Domain\EssentialFrame.Domain\EssentialFrame.Domain.csproj"/>
-    </ItemGroup>
 
+        <ProjectReference Include="..\..\Src\Domain\EssentialFrame.Domain\EssentialFrame.Domain.csproj"/>
+
+
+    </ItemGroup>
 
 
 </Project>
