+++ Src/Domain/EssentialFrame.Domain/Aggregates/IAggregateRoot.cs
@@ -1,0 +1,6 @@
+﻿namespace EssentialFrame.Domain.Aggregates;
+
+public interface IAggregateRoot
+{
+    public Guid AggregateIdentifier { get; }
+}
