+++ Src/Domain/EssentialFrame.Domain/Entities/IEntity.cs
@@ -1,0 +1,6 @@
+﻿namespace EssentialFrame.Domain.Entities;
+
+public interface IEntity
+{
+    public Guid EntityIdentifier { get; }
+}
