+++ Src/Domain/EssentialFrame.Domain/EssentialFrame.Domain.csproj
@@ -29,4 +29,5 @@
         <ProjectReference Include="..\..\Core\EssentialFrame\EssentialFrame.csproj"/>
     </ItemGroup>
 
+
 </Project>
