+++ Src/Domain/EssentialFrame.Domain/Shared/IDeletableObject.cs
@@ -1,0 +1,12 @@
+﻿namespace EssentialFrame.Domain.Shared;
+
+public interface IDeletableObject
+{
+    public DateTimeOffset? DeletedDate { get; }
+
+    public bool IsDeleted { get; }
+
+    public void SafeDelete();
+
+    public void UnDelete();
+}
