+++ Src/Domain/EssentialFrame.Domain/ValueObjects/TypedIdBase.cs
@@ -1,0 +1,16 @@
+﻿namespace EssentialFrame.Domain.ValueObjects;
+
+public class TypedIdBase<T> : ValueObject
+{
+    protected TypedIdBase(T id)
+    {
+        Id = id;
+    }
+
+    public T Id { get; }
+
+    protected override IEnumerable<object> GetEqualityComponents()
+    {
+        yield return Id;
+    }
+}
