+++ Src/Domain/Tests/EssentialFrame.Domain.Tests.UnitTests/EssentialFrame.Domain.Tests.UnitTests.csproj
@@ -32,4 +32,5 @@
         <ProjectReference Include="..\..\EssentialFrame.Domain\EssentialFrame.Domain.csproj"/>
     </ItemGroup>
 
+
 </Project>
