+++ Src/DomainEvents/EssentialFrame.Domain.Events/Core/Snapshots/Interfaces/ISnapshotStrategy.cs
@@ -1,4 +1,4 @@
-﻿using EssentialFrame.Domain.Aggregates;
+﻿using EssentialFrame.Domain.Events.Core.Aggregates;
 
 namespace EssentialFrame.Domain.Events.Core.Snapshots.Interfaces;
