+++ Src/DomainEvents/EssentialFrame.Domain.Events/EssentialFrame.Domain.Events.csproj
@@ -35,4 +35,5 @@
     </ItemGroup>
 
 
+
 </Project>
