+++ Src/DomainEvents/EssentialFrame.Domain.Events/Persistence/Aggregates/Mappers/AggregateMapper.cs
@@ -1,6 +1,8 @@
 ﻿using EssentialFrame.Domain.Aggregates;
+using EssentialFrame.Domain.Events.Core.Aggregates;
 using EssentialFrame.Domain.Events.Persistence.Aggregates.Mappers.Interfaces;
 using EssentialFrame.Domain.Events.Persistence.Aggregates.Models;
+using AggregateRoot = EssentialFrame.Domain.Events.Core.Aggregates.AggregateRoot;
 
 namespace EssentialFrame.Domain.Events.Persistence.Aggregates.Mappers;
