+++ Src/DomainEvents/EssentialFrame.Domain.Events/Persistence/Aggregates/Mappers/Interfaces/IAggregateMapper.cs
@@ -1,5 +1,7 @@
 ﻿using EssentialFrame.Domain.Aggregates;
+using EssentialFrame.Domain.Events.Core.Aggregates;
 using EssentialFrame.Domain.Events.Persistence.Aggregates.Models;
+using AggregateRoot = EssentialFrame.Domain.Events.Core.Aggregates.AggregateRoot;
 
 namespace EssentialFrame.Domain.Events.Persistence.Aggregates.Mappers.Interfaces;
