+++ Src/DomainEvents/EssentialFrame.Domain.Events/Persistence/Aggregates/Services/Interfaces/IAggregateRepository.cs
@@ -1,4 +1,4 @@
-﻿using EssentialFrame.Domain.Aggregates;
+﻿using EssentialFrame.Domain.Events.Core.Aggregates;
 
 namespace EssentialFrame.Domain.Events.Persistence.Aggregates.Services.Interfaces;
