+++ Src/DomainEvents/EssentialFrame.Domain.Events/Persistence/Snapshots/Mappers/Interfaces/ISnapshotMapper.cs
@@ -1,5 +1,5 @@
-﻿using EssentialFrame.Domain.Events.Persistence.Snapshots.Models;
-using EssentialFrame.Domain.Snapshots;
+﻿using EssentialFrame.Domain.Events.Core.Snapshots;
+using EssentialFrame.Domain.Events.Persistence.Snapshots.Models;
 
 namespace EssentialFrame.Domain.Events.Persistence.Snapshots.Mappers.Interfaces;
