+++ Src/DomainEvents/EssentialFrame.Domain.Events/Persistence/Snapshots/Mappers/SnapshotMapper.cs
@@ -1,6 +1,6 @@
-﻿using EssentialFrame.Domain.Events.Persistence.Snapshots.Mappers.Interfaces;
+﻿using EssentialFrame.Domain.Events.Core.Snapshots;
+using EssentialFrame.Domain.Events.Persistence.Snapshots.Mappers.Interfaces;
 using EssentialFrame.Domain.Events.Persistence.Snapshots.Models;
-using EssentialFrame.Domain.Snapshots;
 
 namespace EssentialFrame.Domain.Events.Persistence.Snapshots.Mappers;
