+++ Src/DomainEvents/Tests/EssentialFrame.Domain.Events.Tests.UnitTests/Persistence/Snapshots/SnapshotRepositoryTests.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Bogus;
 using EssentialFrame.Cache.Interfaces;
-using EssentialFrame.Domain.Aggregates;
+using EssentialFrame.Domain.Events.Core.Aggregates;
+using EssentialFrame.Domain.Events.Core.Factories;
+using EssentialFrame.Domain.Events.Core.Snapshots;
 using EssentialFrame.Domain.Events.Core.Snapshots.Interfaces;
 using EssentialFrame.Domain.Events.Persistence.Aggregates.Mappers;
 using EssentialFrame.Domain.Events.Persistence.Aggregates.Mappers.Interfaces;
@@ -14,8 +16,6 @@ using EssentialFrame.Domain.Events.Persistence.Snapshots.Mappers.Interfaces;
 using EssentialFrame.Domain.Events.Persistence.Snapshots.Models;
 using EssentialFrame.Domain.Events.Persistence.Snapshots.Services;
 using EssentialFrame.Domain.Events.Persistence.Snapshots.Services.Interfaces;
-using EssentialFrame.Domain.Factories;
-using EssentialFrame.Domain.Snapshots;
 using EssentialFrame.ExampleApp.Application.Identity;
 using EssentialFrame.ExampleApp.Domain.Posts.Aggregates;
 using EssentialFrame.ExampleApp.Domain.Posts.DomainEvents;
